Inserting comment data into database and remiding exceptions.

+++ src/main/java/chaoimi/community/controller/QuestionController.java
@@ -15,9 +15,10 @@ public class QuestionController {
     private QuestionService questionService;
 
     @GetMapping("/question/{id}")
-    public String question(@PathVariable(name = "id") Integer id,
+    public String question(@PathVariable(name = "id") Long id,
                            Model model){
         QuestionDTO questionDTO = questionService.getById(id);
+        // 增加阅读数
         questionService.incView(id);
         model.addAttribute("question",questionDTO);
         return "question";
